// src/types/apiData/bookings/reservation.ts
export type reservations = {
  id: number;
  booking_number: string;
  shortlet_id: number;
  user_id: number;
  check_in_date: string;
  check_out_date: string;
  check_in_time: string;
  check_out_time: string;
  number_of_days: number;
  number_of_guests: number;
  guest_breakdown: null;
  currency: string;
  caution_fee: number;
  tax_fee: number;
  base_price: number;
  discount_fee: null;
  total_price: number;
  exchange_rate: string;
  caution_fee_NGN: number;
  tax_fee_NGN: number;
  base_price_NGN: number;
  discount_amount_NGN: null;
  total_price_NGN: number;
  status: string;
  payment_status: string;
  payment_method: null;
  promo_code: null;
  notes: null;
  channel: null;
  payment_expires_at: null;
  is_deleted: number;
  deleted_at: null;
  created_at: string;
  updated_at: string;
  transferred_from: null;
  transferred_to: null;
  visitor_code: null;
  account_name: null;
  account_number: null;
  bank_name: null;
  has_requested_refund: number;
  shortlet: {
    id: number;
    name: string;
  };
  user: {
    id: number;
    first_name: string;
    last_name: string;
  };
};